refactor: Test class is relocated
